Move Float trait and ArgminParam type to new file
